requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
openai==1.3.0
langchain==0.0.340
langchain-openai==0.0.2
python-multipart==0.0.6
aiofiles==23.2.1
pyyaml==6.0.1
jinja2==3.1.2
markdown==3.5.1
nltk==3.8.1
networkx==3.2.1
pandas==2.1.3
numpy==1.25.2
matplotlib==3.8.2
seaborn==0.13.0
scikit-learn==1.3.2
asyncio==3.4.3
logging==0.4.9.6
